DRC-2361: updated interface to take in a long for the time instead of an int

+++ DarpaRoboticsChallenge/src/us/ihmc/darpaRoboticsChallenge/maxwellPro/MaxwellProBlackoutScheduleFileGenerator.java
@@ -27,7 +27,7 @@ public class MaxwellProBlackoutScheduleFileGenerator
       {
          String output = time.toString() + ",";
 
-         int blackoutLength = blackoutGenerator.calculateNextBlackoutLength(time.inSeconds());
+         long blackoutLength = blackoutGenerator.calculateNextBlackoutLength(time.inSeconds());
          
          output += blackoutLength + ",";
          time.addSeconds(1);
